refactor: move error-handler to cli/error-handling/
Move error-handler.ts into its own directory as we are going to
add further error handling utils and files.

This is a mechanical refactor.
